+++ src/component/preview/DatasetPreview.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import db from '@db'
   import { tab_selected } from '@lib/store'
-  import Preview_manager from '@lib/preview-manager'
+  import PreviewManager from '@lib/preview-manager'
   import Datatable from '@datatable/Datatable.svelte'
   import Loading from '@frame/Loading.svelte'
 
@@ -14,16 +14,16 @@
     $tab_selected.nb = 0
     if (typeof dataset_preview !== 'string') {
       dataset_data = dataset_preview
-      dataset_data = Preview_manager.add_position(dataset_data)
-      columns = Preview_manager.get_columns(dataset_data)
+      dataset_data = PreviewManager.add_position(dataset_data)
+      columns = PreviewManager.get_columns(dataset_data)
       $tab_selected.nb = dataset_data.length
       return
     }
 
     if (db.preview === undefined) db.preview = {}
     if (!(dataset_preview in db.preview)) {
-      dataset_data = await Preview_manager.load(dataset_preview)
-      Preview_manager.clean_keys(dataset_data)
+      dataset_data = await PreviewManager.load(dataset_preview)
+      PreviewManager.clean_keys(dataset_data)
       db.preview[dataset_preview] = [...dataset_data]
     }
     dataset_data = db.preview[dataset_preview] as any[]
@@ -31,7 +31,7 @@
       delete obj._row_num
       return obj
     })
-    columns = Preview_manager.get_columns(dataset_data)
+    columns = PreviewManager.get_columns(dataset_data)
     $tab_selected.nb = dataset_data.length
   }