+++ src/page/Search.svelte
@@ -2,7 +2,7 @@
   import { onMount } from 'svelte'
   import db from '@db'
   import { search_value, page_content_loaded } from '@lib/store'
-  import { url_param } from '@lib/url-param'
+  import { UrlParam } from '@lib/url-param'
   import Head from '@frame/Head.svelte'
   import Loading from '@frame/Loading.svelte'
   import Tabs from '@tab/Tabs.svelte'
@@ -54,12 +54,12 @@
   }
 
   async function search_input_change() {
-    const url_search_value = url_param.get('search')
+    const url_search_value = UrlParam.get('search')
     if (url_search_value !== $search_value) {
-      url_param.set('search', $search_value)
+      UrlParam.set('search', $search_value)
     }
     if ($search_value === '') {
-      url_param.delete('search')
+      UrlParam.delete('search')
       init_search_recent()
       return false
     }
@@ -95,7 +95,7 @@
 
   let is_empty_input = $derived(['', undefined, null].includes($search_value))
 
-  const url_search_value = url_param.get('search')
+  const url_search_value = UrlParam.get('search')
   if (url_search_value !== false && url_search_value !== '') {
     $search_value = url_search_value
   }